+++ src/features/legal/index.js
@@ -1,3 +1,5 @@
 export { default as LegalPageLayout } from './ui/LegalPageLayout';
 export { default as PrivacyPolicyContent } from './ui/PrivacyPolicyContent';
 export { default as TermsOfServiceContent } from './ui/TermsOfServiceContent';
+export { resolveLegalBackNav } from './model/legalBackNav';
+export { peekAuthDoorResume } from './model/peekAuthDoorResume';